Minor bug fix

+++ client/telegram/src/main/kotlin/com/telegram/bot/handler/BotStateHandler.kt
@@ -40,16 +40,21 @@ enum class BotStateHandler {
             when (telegramBotStateDTO.command) {
                 BotCommandHandler.CANCEL -> return cleanup(user, telegramBotStateDTO)
                 BotCommandHandler.FRIEND_INFO -> {
-                    val friend = userRequestService.getFriendByName(telegramBotStateDTO.id, message.text)
                     telegramBotStateDTO.state = EXPECTING_COMMAND
-                    val stringBuilder = StringBuilder()
-                    for ((name, value) in friend.attributes) {
-                        stringBuilder.append("\t$name: $value\n")
+                    return try {
+                        val friend = userRequestService.getFriendByName(user.id, message.text)
+                        val stringBuilder = StringBuilder()
+                        for ((name, value) in friend.attributes) {
+                            stringBuilder.append("\t$name: $value\n")
+                        }
+                        "Friend - ${friend.name}:\n${stringBuilder}"
+                    } catch (e: RuntimeException) {
+                        "There is no friend with name ${message.text} in your friends list"
                     }
-                     return "Friend - ${friend.name}:\n${stringBuilder}"
                 }
                 BotCommandHandler.ADD_FRIEND -> {
                     telegramBotStateDTO.state = EXPECTING_COMMAND
+                    if (message.text[0] == '/') return "Friend name can not start with /"
                     return try {
                         userRequestService.addFriend(telegramBotStateDTO.id, FriendDTO(0, message.text, emptyMap<String, String?>().toMutableMap()))
                         "Friend was successfully added"
@@ -58,16 +63,25 @@ enum class BotStateHandler {
                     }
                 }
                 BotCommandHandler.REMOVE_FRIEND -> {
-                    val friendId = userRequestService.getFriendByName(telegramBotStateDTO.id, message.text).id
-                    userRequestService.removeFriend(telegramBotStateDTO.id, friendId)
                     telegramBotStateDTO.state = EXPECTING_COMMAND
-                    return "Friend was successfully removed"
+                    return try {
+                        val friendId = userRequestService.getFriendByName(telegramBotStateDTO.id, message.text).id
+                        userRequestService.removeFriend(telegramBotStateDTO.id, friendId)
+                        "Friend was successfully removed"
+                    } catch (e: RuntimeException) {
+                        "There is no friend with name ${message.text} in your friends list"
+                    }
                 }
                 BotCommandHandler.ADD_FRIENDS_ATTRIBUTE, BotCommandHandler.UPDATE_FRIENDS_ATTRIBUTE, BotCommandHandler.REMOVE_FRIENDS_ATTRIBUTE -> {
-                    val friend = userRequestService.getFriendByName(telegramBotStateDTO.id, message.text).id
-                    telegramBotStateDTO.storage["Friend id"] = friend.toString()
-                    telegramBotStateDTO.state = EXPECTING_ATTRIBUTE_NAME
-                    return "Please enter attribute name"
+                    return try {
+                        val friend = userRequestService.getFriendByName(telegramBotStateDTO.id, message.text).id
+                        telegramBotStateDTO.storage["Friend id"] = friend.toString()
+                        telegramBotStateDTO.state = EXPECTING_ATTRIBUTE_NAME
+                        "Please enter attribute name"
+                    } catch (e: RuntimeException) {
+                        telegramBotStateDTO.state = EXPECTING_COMMAND
+                        "There is no friend with name ${message.text} in your friends list"
+                    }
                 }
                 else -> {
                     telegramBotStateDTO.state = EXPECTING_COMMAND
@@ -86,29 +100,46 @@ enum class BotStateHandler {
         ): String {
             when (telegramBotStateDTO.command) {
                 BotCommandHandler.CANCEL -> return cleanup(user, telegramBotStateDTO)
-                BotCommandHandler.ADD_FRIENDS_ATTRIBUTE, BotCommandHandler.UPDATE_FRIENDS_ATTRIBUTE -> {
+                BotCommandHandler.ADD_FRIENDS_ATTRIBUTE -> {
                     telegramBotStateDTO.storage["Attribute name"] = message.text
+                    if (friendRequestService.hasAttribute(telegramBotStateDTO.storage["Friend id"]!!.toLong(), message.text)) {
+                        telegramBotStateDTO.state = EXPECTING_COMMAND
+                        return "Friend already has attribute with name ${message.text}. If you want to change it's value, please use /update_attribute command"
+                    }
                     telegramBotStateDTO.state = EXPECTING_ATTRIBUTE_VALUE
                     return "Please specify its value"
                 }
-                BotCommandHandler.REMOVE_FRIENDS_ATTRIBUTE -> {
-                    if (!telegramBotStateDTO.storage.containsKey("Friend id")) {
-                        telegramBotStateDTO.storage.clear()
+                BotCommandHandler.UPDATE_FRIENDS_ATTRIBUTE -> {
+                    telegramBotStateDTO.storage["Attribute name"] = message.text
+                    if (!friendRequestService.hasAttribute(telegramBotStateDTO.storage["Friend id"]!!.toLong(), message.text)) {
                         telegramBotStateDTO.state = EXPECTING_COMMAND
-                        return "Friend was not specified"
+                        return "Friend doesn't have attribute with name ${message.text}. If you want to add it, please use /add_attribute command"
+                    }
+                    telegramBotStateDTO.state = EXPECTING_ATTRIBUTE_VALUE
+                    return "Please specify its new value"
+                }
+                BotCommandHandler.REMOVE_FRIENDS_ATTRIBUTE -> {
+                    telegramBotStateDTO.state = EXPECTING_COMMAND
+                    if (!friendRequestService.hasAttribute(telegramBotStateDTO.storage["Friend id"]!!.toLong(), message.text)) {
+                        return "Friend doesn't have attribute with name ${message.text}."
                     }
                     friendRequestService.deleteAttribute(telegramBotStateDTO.storage["Friend id"]!!.toLong(), message.text)
-                    telegramBotStateDTO.state = EXPECTING_COMMAND
                     return "Attribute ${message.text} was successfully removed"
                 }
                 BotCommandHandler.ADD_GENERAL_ATTRIBUTE -> {
-                    userRequestService.addGeneralAttribute(telegramBotStateDTO.id, message.text)
                     telegramBotStateDTO.state = EXPECTING_COMMAND
+                    if (userRequestService.hasGeneralAttribute(telegramBotStateDTO.id, message.text)) {
+                        return "You already have general attribute with name ${message.text}"
+                    }
+                    userRequestService.addGeneralAttribute(telegramBotStateDTO.id, message.text)
                     return "Attribute was successfully added"
                 }
                 BotCommandHandler.REMOVE_GENERAL_ATTRIBUTE -> {
-                    userRequestService.removeGeneralAttribute(telegramBotStateDTO.id, message.text)
                     telegramBotStateDTO.state = EXPECTING_COMMAND
+                    if (!userRequestService.hasGeneralAttribute(telegramBotStateDTO.id, message.text)) {
+                        return "You don't have attribute with name ${message.text}."
+                    }
+                    userRequestService.removeGeneralAttribute(telegramBotStateDTO.id, message.text)
                     return "Attribute ${message.text} was successfully removed"
                 }
                 else -> {
@@ -126,25 +157,19 @@ enum class BotStateHandler {
             userRequestService: UserRequestService,
             friendRequestService: FriendRequestService
         ): String {
-            when (telegramBotStateDTO.command) {
-                BotCommandHandler.CANCEL -> return cleanup(user, telegramBotStateDTO)
+            return when (telegramBotStateDTO.command) {
+                BotCommandHandler.CANCEL -> cleanup(user, telegramBotStateDTO)
                 BotCommandHandler.ADD_FRIENDS_ATTRIBUTE, BotCommandHandler.UPDATE_FRIENDS_ATTRIBUTE -> {
-                    if (!telegramBotStateDTO.storage.containsKey("Friend id")
-                        || !telegramBotStateDTO.storage.containsKey("Attribute name")
-                    ) {
-                        telegramBotStateDTO.state = EXPECTING_COMMAND
-                        return "Friend or attribute was not specified"
-                    }
                     friendRequestService.addAttribute(
                         telegramBotStateDTO.storage["Friend id"]!!.toLong(),
                         AttributeDTO(telegramBotStateDTO.storage["Attribute name"]!!, message.text)
                     )
                     telegramBotStateDTO.state = EXPECTING_COMMAND
-                    return "Attribute was successfully added/updated"
+                    "Attribute was successfully added/updated"
                 }
                 else -> {
                     telegramBotStateDTO.state = EXPECTING_COMMAND
-                    return "Invalid command while expecting attribute value"
+                    "Invalid command while expecting attribute value"
                 }
             }
         }
